requirements.txt
python-dotenv==1.0.1
openai==1.79.0
